// projects/shell/src/app/shared/models/common/cifinquiry.model.ts
export interface ContactDetails {
  phoneNumbers: PhoneNumber[];
  emailAddresses: EmailAdresses[];
  addresses: Address[];
}

export interface Address {
  addressId: string;
  county: string;
  constituency: string;
  district: string;
  division: string;
  location: string;
  subLocation: string;
  postalAddress: string;
  cityCode: string;
  city: string;
  village: string;
  estate: string;
  poBox: string;
  postalCode: string;
  stateProvince: string;
  addressType: string;
  preferred: boolean;
  building: string;
  currentPlaceOfResidence: string;
  addressStartDate: string;
  toBeDeleted: boolean;
}

export interface PhoneNumber {
  id: string;
  cityCode: string;
  countryCode: string;
  number: string;
  comment: string;
  phoneType: string;
  preferred: boolean;
  toBeDeleted: boolean;
  code: string;
}

// Generated by https://quicktype.io

export interface EmailAdresses {
  id: string;
  emailAddress: string;
  emailType: string;
  comment: null;
  preferred: boolean;
  toBeDeleted: boolean;
}